Implement cas shortcuts in mvcc store
- put, putIfAbsent
- delete, delete(k, v), deleteRange()
- cas put and delete based on version: vPut, vDelete
- cas put and delete based on revision: rPut, rDelete
- get and range

+++ distributedlog-io/statestore/src/main/java/org/apache/distributedlog/statestore/api/mvcc/op/RangeOp.java
@@ -40,7 +40,7 @@ public interface RangeOp<K, V> extends Op<K, V> {
      */
     boolean isRangeOp();
 
-    int limit();
+    long limit();
 
     long minModRev();
